# api-docs-en.yaml
openapi: 3.0.3
info:
  title: Go-Proxy Configuration API
  description: |
    REST API to manage Go-Proxy load balancer configuration.
    
    **Features:**
    - Dynamic backend server management
    - Automatic scaling actions
    - API Key authentication
    - Min/max limits validation
    
  version: 1.0.0
  contact:
    name: Go-Proxy Team
    url: https://github.com/juanbautista0/go-proxy
servers:
  - url: http://localhost:8082
    description: Local configuration server
security:
  - ApiKeyAuth: []
paths:
...
  /servers:
    post:
      summary: Add backend server
      description: Adds a new server to specified backend
      tags:
        - Servers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddServerRequest'
      responses:
        '201':
          description: Server added successfully
        '400':
          description: Maximum server limit reached or invalid data
        '401':
          description: API Key required or invalid
        '404':
          description: Backend not found
        '500':
          description: Internal server error
    
    put:
      summary: Update existing server
      description: Modifies configuration of existing server
      tags:
        - Servers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateServerRequest'
      responses:
        '200':
          description: Server updated successfully
        '400':
          description: Invalid data
        '401':
          description: API Key required or invalid
        '404':
          description: Server not found
        '500':
          description: Internal server error
    
    delete:
      summary: Remove backend server
      description: Removes server from specified backend
      tags:
        - Servers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RemoveServerRequest'
      responses:
        '200':
          description: Server removed successfully
        '400':
          description: Minimum server limit reached
        '401':
          description: API Key required or invalid
        '404':
          description: Server not found
        '500':
          description: Internal server error
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
      description: |
        API Key for authentication. Contact administrator to get a valid key.
    AdminApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
      description: |
        Admin API Key for administrative operations. Only administrators have access.
  schemas:
...
    AddServerRequest:
      type: object
      required:
        - backend_name
        - url
        - weight
      properties:
        backend_name:
          type: string
          example: "web-servers"
        url:
          type: string
          format: uri
          example: "http://localhost:3004"
        weight:
          type: integer
          minimum: 1
          example: 2
        max_connections:
          type: integer
          minimum: 1
          example: 150
        health_check_endpoint:
          type: string
          example: "/health"
    UpdateServerRequest:
      type: object
      required:
        - backend_name
        - old_url
        - url
        - weight
      properties:
        backend_name:
          type: string
          example: "web-servers"
        old_url:
          type: string
          format: uri
          example: "http://localhost:3004"
        url:
          type: string
          format: uri
          example: "http://localhost:3005"
        weight:
          type: integer
          minimum: 1
          example: 3
        max_connections:
          type: integer
          minimum: 1
          example: 200
        health_check_endpoint:
          type: string
          example: "/status"
    RemoveServerRequest:
      type: object
      required:
        - backend_name
        - server_url
      properties:
        backend_name:
          type: string
          example: "web-servers"
        server_url:
          type: string
          format: uri
          example: "http://localhost:3004"
...